Menu items now contain analyzer options embedded in QVariant.
This is later used to pass options to created analyzer.
Menu is still work in progress.

+++ src/modes/journald/journaldConfiguration.cpp
@@ -67,7 +67,7 @@ QList<JournaldConfiguration::RemoteJournalAddress> JournaldConfiguration::remote
     for (const QString addressItem : m_remoteJournals) {
         RemoteJournalAddress addressInfo;
         addressInfo.address = addressItem.section("|", 0, 0);
-        addressInfo.port = addressItem.section("|", 1);
+        addressInfo.port = addressItem.section("|", 1).toUInt();
         journals.append(addressInfo);
     }
     return journals;
